Drop bi-directional relationship implementations. OBS: requires updated sbt-molecule sbt-molecule project: `git pull`, `sbt publishLocal` molecule project: re-generate boilerplate with `sbt clean compile -Dmolecule=true`
Decided to drop implementations of bidirectional relationships in molecule.

In Datomic relationships can be queried in reverse with for instance some rules:

// Directional query
[:find  ?b ?d
 :where [?a :A/i ?b]
        [?a :A/a ?c]
        [?c :A/i ?d]]

// Bi-directional query
[:find  ?b ?d
 :in    $ %
 :where [?a :A/i ?b]
        (rule?a ?a ?c)
        [?c :A/i ?d]]

// rules:
[
  [(rule?a ?a ?c) [?a :A/a ?c]]
  [(rule?a ?a ?c) [?c :A/a ?a]]
]

Datomic makes this is quite easy to do.

In SQL it quickly becomes unwieldy since we would need to make unions between each pair of relationships, joined in each direction:

SELECT DISTINCT
  A.i   c_1,
  A_a.i c_2
FROM A
INNER JOIN A AS A_a ON A.a = A_a.id
UNION
SELECT DISTINCT
  A_a.i c_1,
  A.i   c_2
FROM A
INNER JOIN A AS A_a ON A.a = A_a.id

When further relationships are added this explodes exponentially since each relationship would need a union of two variations on the query and we get 2 x 2 = 4 unions etc.

To keep things focused I therefore decided to leave bidirectinal queries out of molecule. One can still fall back to make a raw query if such need arises.